+++ frontend/src/locales/bg_BG/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Хумор",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Всички резолюции",
-  "cover-filter-source-all": "Всички източници",
   "cover-filter-style-all": "Всички стилове",
   "cover-filter-uploader-all": "Всички качили",
   "cover-search-failed": "Търсенето на корица е неуспешно: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Видео (стандартизирано)",
   "metadata": "Метаданни",
   "metadata-ids": "Метаданни ID-та",
+  "metadata-providers": "Доставчици на метаданни",
   "metadata-sources-label": "Източници на метаданни",
   "metric-completion": "Завършеност",
   "metric-difficulty": "Трудност",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements ID",
   "provider-screenscraper-id": "ScreenScraper ID",
   "provider-steamgriddb-id": "SteamGridDB ID",
-  "providers-covers": "Други източници",
   "public": "Публично",
   "public-notes": "Публични бележки",
   "qr-scan-to-download": "Сканирай, за да свалиш",
