+++ frontend/src/locales/cs_CZ/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Humor",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Všechna rozlišení",
-  "cover-filter-source-all": "Všechny zdroje",
   "cover-filter-style-all": "Všechny styly",
   "cover-filter-uploader-all": "Všichni nahrávající",
   "cover-search-failed": "Hledání obalu selhalo: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Video (standardizované)",
   "metadata": "Metadata",
   "metadata-ids": "ID metadat",
+  "metadata-providers": "Poskytovatelé metadat",
   "metadata-sources-label": "Zdroje metadat",
   "metric-completion": "Dokončení",
   "metric-difficulty": "Obtížnost",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements ID",
   "provider-screenscraper-id": "ScreenScraper ID",
   "provider-steamgriddb-id": "SteamGridDB ID",
-  "providers-covers": "Další poskytovatelé",
   "public": "Veřejné",
   "public-notes": "Veřejné poznámky",
   "qr-scan-to-download": "Naskenujte pro stažení",
