+++ frontend/src/locales/de_DE/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Humor",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Alle Auflösungen",
-  "cover-filter-source-all": "Alle Quellen",
   "cover-filter-style-all": "Alle Stile",
   "cover-filter-uploader-all": "Alle Uploader",
   "cover-search-failed": "Cover-Suche fehlgeschlagen: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Video (standardisiert)",
   "metadata": "Metadaten",
   "metadata-ids": "Metadaten-IDs",
+  "metadata-providers": "Metadaten-Anbieter",
   "metadata-sources-label": "Metadaten-Quellen",
   "metric-completion": "Abschluss",
   "metric-difficulty": "Schwierigkeit",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements-ID",
   "provider-screenscraper-id": "ScreenScraper-ID",
   "provider-steamgriddb-id": "SteamGridDB-ID",
-  "providers-covers": "Andere Anbieter",
   "public": "Öffentlich",
   "public-notes": "Öffentliche Notizen",
   "qr-scan-to-download": "Scannen zum Herunterladen",
