+++ frontend/src/locales/en_US/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Humor",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "All resolutions",
-  "cover-filter-source-all": "All sources",
   "cover-filter-style-all": "All styles",
   "cover-filter-uploader-all": "All uploaders",
   "cover-search-failed": "Cover search failed: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Video (standardized)",
   "metadata": "Metadata",
   "metadata-ids": "Metadata IDs",
+  "metadata-providers": "Metadata providers",
   "metadata-sources-label": "Metadata sources",
   "metric-completion": "Completion",
   "metric-difficulty": "Difficulty",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements ID",
   "provider-screenscraper-id": "ScreenScraper ID",
   "provider-steamgriddb-id": "SteamGridDB ID",
-  "providers-covers": "Other providers",
   "public": "Public",
   "public-notes": "Public notes",
   "qr-scan-to-download": "Scan to download",
