+++ frontend/src/locales/es_ES/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Humor",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Todas las resoluciones",
-  "cover-filter-source-all": "Todas las fuentes",
   "cover-filter-style-all": "Todos los estilos",
   "cover-filter-uploader-all": "Todos los autores",
   "cover-search-failed": "Búsqueda de portada fallida",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Vídeo (estandarizado)",
   "metadata": "Metadatos",
   "metadata-ids": "IDs de metadatos",
+  "metadata-providers": "Proveedores de metadatos",
   "metadata-sources-label": "Fuentes de metadatos",
   "metric-completion": "Completado",
   "metric-difficulty": "Dificultad",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "ID de RetroAchievements",
   "provider-screenscraper-id": "ID de ScreenScraper",
   "provider-steamgriddb-id": "ID de SteamGridDB",
-  "providers-covers": "Otros proveedores",
   "public": "Pública",
   "public-notes": "Notas públicas",
   "qr-scan-to-download": "Escanear para descargar",
