+++ frontend/src/locales/fr_FR/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Humour",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Toutes les résolutions",
-  "cover-filter-source-all": "Toutes les sources",
   "cover-filter-style-all": "Tous les styles",
   "cover-filter-uploader-all": "Tous les contributeurs",
   "cover-search-failed": "Échec de la recherche de jaquette : {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Vidéo (standardisée)",
   "metadata": "Métadonnées",
   "metadata-ids": "IDs de métadonnées",
+  "metadata-providers": "Fournisseurs de métadonnées",
   "metadata-sources-label": "Sources de métadonnées",
   "metric-completion": "Complétion",
   "metric-difficulty": "Difficulté",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "ID RetroAchievements",
   "provider-screenscraper-id": "ID ScreenScraper",
   "provider-steamgriddb-id": "ID SteamGridDB",
-  "providers-covers": "Autres fournisseurs",
   "public": "Publique",
   "public-notes": "Notes publiques",
   "qr-scan-to-download": "Scanner pour télécharger",
