+++ frontend/src/locales/hu_HU/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Humor",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Összes felbontás",
-  "cover-filter-source-all": "Összes forrás",
   "cover-filter-style-all": "Összes stílus",
   "cover-filter-uploader-all": "Összes feltöltő",
   "cover-search-failed": "A borítókeresés sikertelen: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Videó (szabványosított)",
   "metadata": "Metaadat",
   "metadata-ids": "Metaadat ID-k",
+  "metadata-providers": "Metaadat-szolgáltatók",
   "metadata-sources-label": "Metaadat-források",
   "metric-completion": "Befejezés",
   "metric-difficulty": "Nehézség",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements azonosító",
   "provider-screenscraper-id": "ScreenScraper azonosító",
   "provider-steamgriddb-id": "SteamGridDB azonosító",
-  "providers-covers": "Más szolgáltatók",
   "public": "Publikus",
   "public-notes": "Publikus jegyzetek",
   "qr-scan-to-download": "Olvasd be a letöltéshez",
