+++ frontend/src/locales/it_IT/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Umorismo",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Tutte le risoluzioni",
-  "cover-filter-source-all": "Tutte le fonti",
   "cover-filter-style-all": "Tutti gli stili",
   "cover-filter-uploader-all": "Tutti gli autori",
   "cover-search-failed": "Ricerca copertina non riuscita: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Video (standardizzato)",
   "metadata": "Metadati",
   "metadata-ids": "ID metadati",
+  "metadata-providers": "Provider di metadati",
   "metadata-sources-label": "Fonti metadati",
   "metric-completion": "Completamento",
   "metric-difficulty": "Difficoltà",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "ID RetroAchievements",
   "provider-screenscraper-id": "ID ScreenScraper",
   "provider-steamgriddb-id": "ID SteamGridDB",
-  "providers-covers": "Altri provider",
   "public": "Pubblica",
   "public-notes": "Note pubbliche",
   "qr-scan-to-download": "Scansiona per scaricare",
