+++ frontend/src/locales/ja_JP/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "ユーモア",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "すべての解像度",
-  "cover-filter-source-all": "すべてのソース",
   "cover-filter-style-all": "すべてのスタイル",
   "cover-filter-uploader-all": "すべての投稿者",
   "cover-search-failed": "カバー検索に失敗しました: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "ビデオ（標準化）",
   "metadata": "メタデータ",
   "metadata-ids": "メタデータID",
+  "metadata-providers": "メタデータプロバイダ",
   "metadata-sources-label": "メタデータソース",
   "metric-completion": "完了度",
   "metric-difficulty": "難易度",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements ID",
   "provider-screenscraper-id": "ScreenScraper ID",
   "provider-steamgriddb-id": "SteamGridDB ID",
-  "providers-covers": "その他のプロバイダ",
   "public": "公開",
   "public-notes": "公開メモ",
   "qr-scan-to-download": "スキャンしてダウンロード",
