+++ frontend/src/locales/ko_KR/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "유머",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "모든 해상도",
-  "cover-filter-source-all": "모든 출처",
   "cover-filter-style-all": "모든 스타일",
   "cover-filter-uploader-all": "모든 업로더",
   "cover-search-failed": "커버 검색 실패: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "비디오 (표준화)",
   "metadata": "메타데이터",
   "metadata-ids": "메타데이터 ID",
+  "metadata-providers": "메타데이터 제공자",
   "metadata-sources-label": "메타데이터 소스",
   "metric-completion": "완료율",
   "metric-difficulty": "난이도",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements ID",
   "provider-screenscraper-id": "ScreenScraper ID",
   "provider-steamgriddb-id": "SteamGridDB ID",
-  "providers-covers": "다른 제공자",
   "public": "공개",
   "public-notes": "공개 노트",
   "qr-scan-to-download": "스캔하여 다운로드",
