+++ frontend/src/locales/pl_PL/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Humor",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Wszystkie rozdzielczości",
-  "cover-filter-source-all": "Wszystkie źródła",
   "cover-filter-style-all": "Wszystkie style",
   "cover-filter-uploader-all": "Wszyscy przesyłający",
   "cover-search-failed": "Wyszukiwanie okładki nie powiodło się: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Wideo (znormalizowane)",
   "metadata": "Metadane",
   "metadata-ids": "ID-y metadanych",
+  "metadata-providers": "Dostawcy metadanych",
   "metadata-sources-label": "Źródła metadanych",
   "metric-completion": "Ukończenie",
   "metric-difficulty": "Trudność",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements ID",
   "provider-screenscraper-id": "ScreenScraper ID",
   "provider-steamgriddb-id": "SteamGridDB ID",
-  "providers-covers": "Inni dostawcy",
   "public": "Publiczna",
   "public-notes": "Notatki publiczne",
   "qr-scan-to-download": "Zeskanuj, aby pobrać",
