+++ frontend/src/locales/pt_BR/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Humor",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Todas as resoluções",
-  "cover-filter-source-all": "Todas as fontes",
   "cover-filter-style-all": "Todos os estilos",
   "cover-filter-uploader-all": "Todos os autores",
   "cover-search-failed": "Falha na busca de capa: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Vídeo (padronizado)",
   "metadata": "Metadados",
   "metadata-ids": "IDs Metadados",
+  "metadata-providers": "Provedores de metadados",
   "metadata-sources-label": "Fontes de metadados",
   "metric-completion": "Conclusão",
   "metric-difficulty": "Dificuldade",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "ID do RetroAchievements",
   "provider-screenscraper-id": "ID do ScreenScraper",
   "provider-steamgriddb-id": "ID do SteamGridDB",
-  "providers-covers": "Outros provedores",
   "public": "Pública",
   "public-notes": "Notas públicas",
   "qr-scan-to-download": "Escaneie para baixar",
