+++ frontend/src/locales/ro_RO/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Umor",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Toate rezoluțiile",
-  "cover-filter-source-all": "Toate sursele",
   "cover-filter-style-all": "Toate stilurile",
   "cover-filter-uploader-all": "Toți încărcătorii",
   "cover-search-failed": "Căutarea coperții a eșuat: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Video (standardizat)",
   "metadata": "Metadate",
   "metadata-ids": "ID-uri Metadate",
+  "metadata-providers": "Furnizori de metadate",
   "metadata-sources-label": "Surse de metadate",
   "metric-completion": "Finalizare",
   "metric-difficulty": "Dificultate",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "ID RetroAchievements",
   "provider-screenscraper-id": "ID ScreenScraper",
   "provider-steamgriddb-id": "ID SteamGridDB",
-  "providers-covers": "Alți furnizori",
   "public": "Publică",
   "public-notes": "Notițe publice",
   "qr-scan-to-download": "Scanează pentru a descărca",
