+++ frontend/src/locales/ru_RU/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Юмор",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Все разрешения",
-  "cover-filter-source-all": "Все источники",
   "cover-filter-style-all": "Все стили",
   "cover-filter-uploader-all": "Все загрузившие",
   "cover-search-failed": "Поиск обложки не удался: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Видео (стандартизированное)",
   "metadata": "Метаданные",
   "metadata-ids": "ID-ы метаданных",
+  "metadata-providers": "Поставщики метаданных",
   "metadata-sources-label": "Источники метаданных",
   "metric-completion": "Прохождение",
   "metric-difficulty": "Сложность",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements ID",
   "provider-screenscraper-id": "ScreenScraper ID",
   "provider-steamgriddb-id": "SteamGridDB ID",
-  "providers-covers": "Другие провайдеры",
   "public": "Публичная",
   "public-notes": "Публичные заметки",
   "qr-scan-to-download": "Отсканируйте для скачивания",
