+++ frontend/src/locales/tr_TR/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "Mizah",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "Tüm çözünürlükler",
-  "cover-filter-source-all": "Tüm kaynaklar",
   "cover-filter-style-all": "Tüm stiller",
   "cover-filter-uploader-all": "Tüm yükleyenler",
   "cover-search-failed": "Kapak araması başarısız: {error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "Video (standartlaştırılmış)",
   "metadata": "Meta veri",
   "metadata-ids": "Meta veri ID'leri",
+  "metadata-providers": "Meta veri sağlayıcıları",
   "metadata-sources-label": "Meta veri kaynakları",
   "metric-completion": "Tamamlama",
   "metric-difficulty": "Zorluk",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements ID",
   "provider-screenscraper-id": "ScreenScraper ID",
   "provider-steamgriddb-id": "SteamGridDB ID",
-  "providers-covers": "Diğer sağlayıcılar",
   "public": "Herkese açık",
   "public-notes": "Herkese açık notlar",
   "qr-scan-to-download": "İndirmek için tarat",
