+++ frontend/src/locales/zh_CN/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "幽默",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "所有分辨率",
-  "cover-filter-source-all": "所有来源",
   "cover-filter-style-all": "所有样式",
   "cover-filter-uploader-all": "所有上传者",
   "cover-search-failed": "封面搜索失败：{error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "视频（标准化）",
   "metadata": "元数据",
   "metadata-ids": "元数据ID",
+  "metadata-providers": "元数据提供商",
   "metadata-sources-label": "元数据来源",
   "metric-completion": "完成度",
   "metric-difficulty": "难度",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements ID",
   "provider-screenscraper-id": "ScreenScraper ID",
   "provider-steamgriddb-id": "SteamGridDB ID",
-  "providers-covers": "其他来源",
   "public": "公开",
   "public-notes": "公共笔记",
   "qr-scan-to-download": "扫码下载",
