+++ frontend/src/locales/zh_TW/rom.json
@@ -70,7 +70,6 @@
   "cover-content-humor": "幽默",
   "cover-content-nsfw": "NSFW",
   "cover-filter-resolution-all": "所有解析度",
-  "cover-filter-source-all": "所有來源",
   "cover-filter-style-all": "所有樣式",
   "cover-filter-uploader-all": "所有上傳者",
   "cover-search-failed": "封面搜尋失敗：{error}",
@@ -217,6 +216,7 @@
   "media-video-normalized": "影片（標準化）",
   "metadata": "元數據",
   "metadata-ids": "元數據ID",
+  "metadata-providers": "中繼資料提供商",
   "metadata-sources-label": "元數據來源",
   "metric-completion": "完成度",
   "metric-difficulty": "難度",
@@ -287,7 +287,6 @@
   "provider-retroachievements-id": "RetroAchievements ID",
   "provider-screenscraper-id": "ScreenScraper ID",
   "provider-steamgriddb-id": "SteamGridDB ID",
-  "providers-covers": "其他提供商",
   "public": "公開",
   "public-notes": "公共筆記",
   "qr-scan-to-download": "掃描以下載",
